Move `invokeAndExecute` into `client.go`
